# crates/bindizr/src/api/dnssec_policy.rs
use axum::{
    Json, Router,
    http::StatusCode,
    response::{IntoResponse, Response},
    routing,
};
use bindizr_service::{
    dnssec_policy::DnssecPolicyService,
    types::{
        CreateDnssecPolicyRequest, DEFAULT_PAGE_LIMIT, DnssecPolicyResponse, ErrorResponse,
        GetDnssecPolicyResponse, MessageResponse, PageFilter, PaginatedResponse,
        UpdateDnssecPolicyRequest,
    },
};
use serde::Deserialize;

use crate::api::{
    RequestCaller,
    error::{ApiError, Path, Query},
    middleware::body_parser::JsonBody,
};

pub(crate) struct DnssecPolicyApi;

impl DnssecPolicyApi {
    /// Build the DNSSEC policy API routes.
    pub(crate) async fn routes() -> Router {
        Router::new()
            .route("/dnssec-policies", routing::get(list_dnssec_policies))
            .route("/dnssec-policies", routing::post(create_dnssec_policy))
            .route("/dnssec-policies/{name}", routing::get(get_dnssec_policy))
            .route(
                "/dnssec-policies/{name}",
                routing::put(update_dnssec_policy),
            )
            .route(
                "/dnssec-policies/{name}",
                routing::delete(delete_dnssec_policy),
            )
    }
}

#[derive(Deserialize)]
pub(crate) struct DnssecPolicyNameParam {
    name: String,
}

/// List all DNSSEC policies.
#[utoipa::path(
        get,
        path = "/dnssec-policies",
        tag = "DNSSEC",
        summary = "List all DNSSEC policies",
        params(PageFilter),
        description = "Lists every DNSSEC policy: the named signing-parameter bundles zones sign under. A `default` policy (ECDSA P-256 CSK, NSEC, 14-day signatures re-signed with 5 days left) is seeded at startup.",
        responses(
            (status = 200, description = "All DNSSEC policies", body = PaginatedResponse<GetDnssecPolicyResponse>),
            (status = 401, description = "Unauthorized", body = ErrorResponse),
            (status = 403, description = "A global API token is required", body = ErrorResponse),
            (status = 500, description = "Internal server error", body = ErrorResponse)
        )
)]
pub(crate) async fn list_dnssec_policies(
    RequestCaller(caller): RequestCaller,
    Query(mut page): Query<PageFilter>,
) -> Result<Response, ApiError> {
    page.limit = page.limit.or(Some(DEFAULT_PAGE_LIMIT));
    let response = DnssecPolicyService::list(&caller, page).await?;
    Ok((StatusCode::OK, Json(response)).into_response())
}

/// Create a DNSSEC policy.
#[utoipa::path(
        post,
        path = "/dnssec-policies",
        tag = "DNSSEC",
        summary = "Create a DNSSEC policy",
        description = "Creates a DNSSEC policy. The algorithm (`ecdsap256sha256` by default; also `ecdsap384sha384`, `ed25519`, `ed448`, `rsasha256`, `rsasha512`), denial mode (`nsec3` by default, or `nsec`), and key layout (`split_keys`) are fixed once created; the timing fields can be edited later. Omitted fields take the built-in defaults.",
        request_body = CreateDnssecPolicyRequest,
        responses(
            (status = 201, description = "DNSSEC policy created successfully", body = DnssecPolicyResponse),
            (status = 400, description = "Bad request, invalid input", body = ErrorResponse),
            (status = 401, description = "Unauthorized", body = ErrorResponse),
            (status = 403, description = "A global API token is required", body = ErrorResponse),
            (status = 409, description = "A DNSSEC policy with the same name already exists", body = ErrorResponse),
            (status = 415, description = "Unsupported media type, expected JSON request body", body = ErrorResponse),
            (status = 500, description = "Internal server error", body = ErrorResponse)
        )
)]
pub(crate) async fn create_dnssec_policy(
    RequestCaller(caller): RequestCaller,
    JsonBody(body): JsonBody<CreateDnssecPolicyRequest>,
) -> Result<Response, ApiError> {
    let policy = DnssecPolicyService::create(&caller, body).await?;
    let response = DnssecPolicyResponse {
        dnssec_policy: GetDnssecPolicyResponse::from_policy(&policy),
    };
    Ok((StatusCode::CREATED, Json(response)).into_response())
}

/// Get one DNSSEC policy by name.
#[utoipa::path(
        get,
        path = "/dnssec-policies/{name}",
        tag = "DNSSEC",
        summary = "Get a specific DNSSEC policy",
        params(
            ("name" = String, Path, description = "The name of the DNSSEC policy.")
        ),
        responses(
            (status = 200, description = "The DNSSEC policy", body = DnssecPolicyResponse),
            (status = 401, description = "Unauthorized", body = ErrorResponse),
            (status = 403, description = "A global API token is required", body = ErrorResponse),
            (status = 404, description = "DNSSEC policy not found", body = ErrorResponse),
            (status = 500, description = "Internal server error", body = ErrorResponse)
        )
)]
pub(crate) async fn get_dnssec_policy(
    RequestCaller(caller): RequestCaller,
    Path(params): Path<DnssecPolicyNameParam>,
) -> Result<Response, ApiError> {
    let policy = DnssecPolicyService::get(&caller, &params.name).await?;
    let response = DnssecPolicyResponse {
        dnssec_policy: GetDnssecPolicyResponse::from_policy(&policy),
    };
    Ok((StatusCode::OK, Json(response)).into_response())
}

/// Edit a DNSSEC policy's timing fields.
#[utoipa::path(
        put,
        path = "/dnssec-policies/{name}",
        tag = "DNSSEC",
        summary = "Edit a DNSSEC policy's timing",
        description = "Edits the policy's signature validity, re-sign threshold, and scheduled ZSK lifetime; an omitted field keeps its value. The algorithm, denial mode, and key layout cannot change: move zones to another policy instead. Zones under the policy pick the new values up on their next signing pass or maintenance scan.",
        params(
            ("name" = String, Path, description = "The name of the DNSSEC policy.")
        ),
        request_body = UpdateDnssecPolicyRequest,
        responses(
            (status = 200, description = "DNSSEC policy updated", body = DnssecPolicyResponse),
            (status = 400, description = "Bad request, invalid input", body = ErrorResponse),
            (status = 401, description = "Unauthorized", body = ErrorResponse),
            (status = 403, description = "A global API token is required", body = ErrorResponse),
            (status = 404, description = "DNSSEC policy not found", body = ErrorResponse),
            (status = 415, description = "Unsupported media type, expected JSON request body", body = ErrorResponse),
            (status = 500, description = "Internal server error", body = ErrorResponse)
        )
)]
pub(crate) async fn update_dnssec_policy(
    RequestCaller(caller): RequestCaller,
    Path(params): Path<DnssecPolicyNameParam>,
    JsonBody(body): JsonBody<UpdateDnssecPolicyRequest>,
) -> Result<Response, ApiError> {
    let policy = DnssecPolicyService::update(&caller, &params.name, body).await?;
    let response = DnssecPolicyResponse {
        dnssec_policy: GetDnssecPolicyResponse::from_policy(&policy),
    };
    Ok((StatusCode::OK, Json(response)).into_response())
}

/// Delete a DNSSEC policy no zone signs under.
#[utoipa::path(
        delete,
        path = "/dnssec-policies/{name}",
        tag = "DNSSEC",
        summary = "Delete a DNSSEC policy",
        description = "Deletes a DNSSEC policy. Refused for the built-in `default` policy and while any zone signs under it.",
        params(
            ("name" = String, Path, description = "The name of the DNSSEC policy.")
        ),
        responses(
            (status = 200, description = "DNSSEC policy deleted successfully", body = MessageResponse),
            (status = 400, description = "The built-in default policy cannot be deleted", body = ErrorResponse),
            (status = 401, description = "Unauthorized", body = ErrorResponse),
            (status = 403, description = "A global API token is required", body = ErrorResponse),
            (status = 404, description = "DNSSEC policy not found", body = ErrorResponse),
            (status = 409, description = "DNSSEC policy is still used by signed zones", body = ErrorResponse),
            (status = 500, description = "Internal server error", body = ErrorResponse)
        )
)]
pub(crate) async fn delete_dnssec_policy(
    RequestCaller(caller): RequestCaller,
    Path(params): Path<DnssecPolicyNameParam>,
) -> Result<Response, ApiError> {
    DnssecPolicyService::delete(&caller, &params.name).await?;
    let response = MessageResponse {
        message: "DNSSEC policy deleted successfully".to_string(),
    };
    Ok((StatusCode::OK, Json(response)).into_response())
}