# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bindizr HTTP API
  description: This is the API documentation for Bindizr
  contact:
    email: kevin136583@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0-beta.7
paths:
  /dnssec-policies:
    get:
      tags:
      - DNSSEC
      summary: List all DNSSEC policies
      description: 'Lists every DNSSEC policy: the named signing-parameter bundles zones sign under. A `default` policy (ECDSA P-256 CSK, NSEC, 14-day signatures re-signed with 5 days left) is seeded at startup.'
      operationId: list_dnssec_policies
      parameters:
      - name: limit
        in: query
        description: Items per page; the HTTP API defaults it, the daemon socket does not.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
      responses:
        '200':
          description: All DNSSEC policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_GetDnssecPolicyResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: A global API token is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - DNSSEC
      summary: Create a DNSSEC policy
      description: Creates a DNSSEC policy. The algorithm (`ecdsap256sha256` by default; also `ecdsap384sha384`, `ed25519`, `ed448`, `rsasha256`, `rsasha512`), denial mode (`nsec3` by default, or `nsec`), and key layout (`split_keys`) are fixed once created; the timing fields can be edited later. Omitted fields take the built-in defaults.
      operationId: create_dnssec_policy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDnssecPolicyRequest'
        required: true
      responses:
        '201':
          description: DNSSEC policy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnssecPolicyResponse'
        '400':
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: A global API token is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A DNSSEC policy with the same name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type, expected JSON request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dnssec-policies/{name}:
    get:
      tags:
      - DNSSEC
      summary: Get a specific DNSSEC policy
      operationId: get_dnssec_policy
      parameters:
      - name: name
        in: path
        description: The name of the DNSSEC policy.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The DNSSEC policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnssecPolicyResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: A global API token is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: DNSSEC policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - DNSSEC
      summary: Edit a DNSSEC policy's timing
      description: 'Edits the policy''s signature validity, re-sign threshold, and scheduled ZSK lifetime; an omitted field keeps its value. The algorithm, denial mode, and key layout cannot change: move zones to another policy instead. Zones under the policy pick the new values up on their next signing pass or maintenance scan.'
      operationId: update_dnssec_policy
      parameters:
      - name: name
        in: path
        description: The name of the DNSSEC policy.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDnssecPolicyRequest'
        required: true
      responses:
        '200':
          description: DNSSEC policy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnssecPolicyResponse'
        '400':
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: A global API token is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: DNSSEC policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type, expected JSON request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - DNSSEC
      summary: Delete a DNSSEC policy
      description: Deletes a DNSSEC policy. Refused for the built-in `default` policy and while any zone signs under it.
      operationId: delete_dnssec_policy
      parameters:
      - name: name
        in: path
        description: The name of the DNSSEC policy.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: DNSSEC policy deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: The built-in default policy cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: A global API token is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: DNSSEC policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: DNSSEC policy is still used by signed zones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateDnssecPolicyRequest:
      type: object
      description: |-
        Request body for creating a DNSSEC policy. The key layout, algorithm,
        and denial mode are fixed once created. Omitted fields use the built-in
        defaults, independent of edits to the policy named `default`.
      required:
      - name
      properties:
        algorithm:
          type:
          - string
          - 'null'
          description: |-
            Defaults to `ecdsap256sha256`; also accepts `ecdsap384sha384`,
            `ed25519`, `ed448`, `rsasha256`, and `rsasha512`.
          example: ecdsap256sha256
        denial:
          type:
          - string
          - 'null'
          description: |-
            Denial-of-existence mode: `nsec3` (default, RFC 9276 parameters) or
            `nsec`, which leaves the zone's names walkable.
          example: nsec3
        name:
          type: string
          example: strict
        signature_refresh_days:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            Re-sign when a signature has fewer than this many days left (default
            5); must be below the validity.
          example: 5
          minimum: 0
        signature_validity_days:
          type:
          - integer
          - 'null'
          format: int32
          description: Days a new signature stays valid (default 14).
          example: 14
          minimum: 0
        split_keys:
          type: boolean
          description: |-
            A KSK/ZSK pair instead of one CSK, so the ZSK rolls without touching
            the parent DS.
          example: false
        zsk_lifetime_days:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            Days an active ZSK may sign before the scheduler rolls it; 0 (the
            default) disables scheduled rolls.
          example: 90
          minimum: 0
...
    DnssecPolicyResponse:
      type: object
      description: A DNSSEC policy wrapped in a response envelope.
      required:
      - dnssec_policy
      properties:
        dnssec_policy:
          $ref: '#/components/schemas/GetDnssecPolicyResponse'
...
    ErrorResponse:
      type: object
      description: 'Generic error response: a plain description plus a machine-readable code.'
      required:
      - error
      - code
      properties:
        code:
          type: string
          example: ZONE_NOT_FOUND
        error:
          type: string
          example: Zone with name 'example.com' not found
...
    GetDnssecPolicyResponse:
      type: object
      description: API representation of a DNSSEC policy.
      required:
      - id
      - name
      - algorithm
      - denial
      - split_keys
      - signature_validity_days
      - signature_refresh_days
      - zsk_lifetime_days
      - created_at
      properties:
        algorithm:
          type: string
          example: ecdsap256sha256
        created_at:
          type: string
          format: date-time
        denial:
          type: string
          description: '`nsec` or `nsec3`.'
          example: nsec3
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: default
        signature_refresh_days:
          type: integer
          format: int32
          example: 5
          minimum: 0
        signature_validity_days:
          type: integer
          format: int32
          example: 14
          minimum: 0
        split_keys:
          type: boolean
          example: false
        zsk_lifetime_days:
          type: integer
          format: int32
          description: 0 disables scheduled ZSK rollovers.
          example: 0
          minimum: 0
...
    MessageResponse:
      type: object
      description: Generic success message response.
      required:
      - message
      properties:
        message:
          type: string
          example: Deleted successfully
    PaginatedResponse_GetDnssecPolicyResponse:
      type: object
      description: A page of items together with its pagination metadata.
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            type: object
            description: API representation of a DNSSEC policy.
            required:
            - id
            - name
            - algorithm
            - denial
            - split_keys
            - signature_validity_days
            - signature_refresh_days
            - zsk_lifetime_days
            - created_at
            properties:
              algorithm:
                type: string
                example: ecdsap256sha256
              created_at:
                type: string
                format: date-time
              denial:
                type: string
                description: '`nsec` or `nsec3`.'
                example: nsec3
              id:
                type: integer
                format: int32
                example: 1
              name:
                type: string
                example: default
              signature_refresh_days:
                type: integer
                format: int32
                example: 5
                minimum: 0
              signature_validity_days:
                type: integer
                format: int32
                example: 14
                minimum: 0
              split_keys:
                type: boolean
                example: false
              zsk_lifetime_days:
                type: integer
                format: int32
                description: 0 disables scheduled ZSK rollovers.
                example: 0
                minimum: 0
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      description: Pagination window and total count for a list response.
      required:
      - limit
      - offset
      - total
      properties:
        limit:
          type: integer
          format: int32
          example: 50
          minimum: 0
        offset:
          type: integer
          format: int64
          example: 0
          minimum: 0
        total:
          type: integer
          format: int64
          example: 125
          minimum: 0
...
    UpdateDnssecPolicyRequest:
      type: object
      description: |-
        Request body for editing a DNSSEC policy's timing; an omitted field keeps
        its value. Takes effect on the next signing pass or maintenance scan.
      properties:
        signature_refresh_days:
          type:
          - integer
          - 'null'
          format: int32
          example: 7
          minimum: 0
        signature_validity_days:
          type:
          - integer
          - 'null'
          format: int32
          example: 30
          minimum: 0
        zsk_lifetime_days:
          type:
          - integer
          - 'null'
          format: int32
          example: 90
          minimum: 0
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...